requirements.txt
pandas
matplotlib
SciencePlots
pyyaml